Update patch set 1
Patch Set 1:

It was changed to 4kb to improve boot speed- on .29 this was the default but it was raised in .32. 4kb seems to work best on NAND flash. I'd suggest a better change might be to set this via the bdi sysfs node for the sdcard only. Probably something we should do for all devices actually.

Patch-set: 1