Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

As I noted on the binary commit, this backport works well for Linux hosts but fails on Windows.  When UMS is chosen, Windows thinks for a moment, and then the phone resets back to UMS disabled.

So, tentative approval, but we'll have to figure this out...

Patch-set: 1