Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Compiles, kernel and flash led device operates as expected, with correct permissions.

Theory is HTC makes the device 666 since it ships a flashlight app with the phone.  Changed permissions on the device to bring in line with other devices.

Patch-set: 1